backend/requirements.txt
fastapi
uvicorn
pydicom
groq
python-multipart
pillow
numpy
python-dotenv
